Update README.md
I just bored in covid and try to play with turtle libriary in python. How it goes in forward backward and how it moves at a fixed angle, I was thougth that to design an image of something different and I was found to design four bond of chemistry so then I was able to do this.